split updates feed to container and component
