+++ static/src/components/updates-feed/index.js
@@ -1,0 +1,1 @@
+export { default as UpdatesFeed } from './updates-feed'
