+++ static/src/components/updates-feed/updates-feed.scss
